requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
comm==0.1.2
contourpy==1.0.7
crypto==1.4.1
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-Script==2.0.6
fonttools==4.38.0
fqdn==1.5.1
gcloud==0.18.3
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jwcrypto==1.4.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
mysqlclient==2.1.1
Naked==0.1.32
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauth2client==4.1.3
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pygame==2.2.0
Pygments==2.14.0
pyparsing==3.0.9
Pyrebase4==4.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-jwt==4.0.0
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
Send2Trash==1.8.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
whitenoise==6.4.0
wrapt==1.15.0
WTForms==3.0.1
